refactoring(nhiss): split register info class
